refactor(checker): split jsdoc heritage mismatch check
